+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added control modes and control times
+- added legend to MultipleLineChart in myems-web
 ### Changed
 - replaced datetime.strftime() with datetime.isoformat() in myems-api
 - changed space reports to automaticly submit in myems-web
 - changed reporting period of space reports to one week in myems-web
+- updated label interval of MultipleLineChart in myems-web
 ### Fixed
 - fixed datetime isoformat issues in myems-api
 - fixed api key datetime issue in myems-api and myems-admin